% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{admin}
\alias{admin}
\title{Timepoint of Rubella vaccination}
\usage{
admin(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with the comments from Table ESTIMATE_REQUIRED
}
\description{
Timepoint of Rubella vaccination
}
\details{
SELECT annum, vaccine, coverage
 FROM REPORTED_COVERAGE
 WHERE country = CCODE AND annum >= 1997 AND coverage > 0 
   AND vaccine IN vaxs AND type = "admin"
}
