% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{country}
\alias{country}
\title{Load country name}
\usage{
country(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Country name
}
\description{
Load country name
}
