% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{decisions}
\alias{decisions}
\title{Obtain list of working group decisions}
\usage{
decisions(mdb, ccode, Survey)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}

\item{Survey}{Survey data (for decisions such as acceptSurvey etc.)}
}
\value{
Decisions in long format
}
\description{
Obtain list of working group decisions
}
\details{
SELECT vaccine, annum, coverage
  FROM ESTIMATED_COVERAGE
  WHERE country = CCODE AND annum = 1997 and vaccine IN vaxs

SELECT *
  FROM SURVEY_COVERAGE
  WHERE surveyId IN Ids AND reanalyzed = "No" AND cohortYear >= 1997
    AND validity = "crude" AND vaccine IN vaxs

SELECT *
  FROM WGD
  WHERE country = CCODE
}
