% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{est_req}
\alias{est_req}
\title{Check list of required estimates}
\usage{
est_req(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table
}
\description{
Check list of required estimates
}
\details{
SELECT vaccine, annum
  FROM ESTIMATE_REQUIRED WHERE country = CCODE AND annum >= 1997
}
