% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuenic_ver_4.1.R
\name{estimate}
\alias{estimate}
\title{Run from command line}
\usage{
estimate(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{command line arguments (default)}
}
\value{
TRUE on success
}
\description{
Run from command line
}
