% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{gov}
\alias{gov}
\title{Coverage data from government}
\usage{
gov(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with reported coverage
}
\description{
Coverage data from government
}
\details{
SELECT annum, vaccine, coverage
 FROM REPORTED_COVERAGE
 WHERE country = CCODE AND annum >= 1997 AND coverage > 0 
   AND vaccine IN vaxs AND type = "official"
}
