% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{legacy}
\alias{legacy}
\title{Legacy estimates (for comparison if database changes)}
\usage{
legacy(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with legacy coverage
}
\description{
Legacy estimates (for comparison if database changes)
}
\details{
SELECT annum, vaccine, coverage
  FROM ESTIMATED_COVERAGE
  WHERE country = ccode AND annum >= 1997 AND coverage > 0 
    AND vaccine IN vaxs
}
