% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_mdb.R
\name{mdb_get}
\alias{mdb_get}
\title{Interface to mdb-tools}
\usage{
mdb_get(mdb, tab)
}
\arguments{
\item{mdb}{Path to MS Access database (extension mdb)}

\item{tab}{Which table should be loaded}
}
\value{
Data frame with the requested table.
}
\description{
Interface to mdb-tools
}
\details{
Under Windows, install mdb-tools from Cygwin and put it into the PATH
}
