% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{rubella}
\alias{rubella}
\title{Timepoint of Rubella vaccination}
\usage{
rubella(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with the comments from Table ESTIMATE_REQUIRED
}
\description{
Timepoint of Rubella vaccination
}
\details{
SELECT vaccine, annum
  FROM ESTIMATE_REQUIRED WHERE country = CCODE AND annum >= 1997
}
