% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{sawtooth}
\alias{sawtooth}
\title{sawtooth: jumps in reported vaccine coverage of more than 10 percentage
points are not plausible}
\usage{
sawtooth()
}
\value{
10
}
\description{
sawtooth: jumps in reported vaccine coverage of more than 10 percentage
points are not plausible
}
