% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{surveys}
\alias{surveys}
\title{Obtain survey data}
\usage{
surveys(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Survey data in long format
}
\description{
Obtain survey data
}
\details{
SELECT surveyId
  FROM SURVEY_DESCRIPTION
  WHERE ISO3 = CCODE

SELECT *
  FROM SURVEY_COVERAGE
  WHERE surveyId IN Ids AND reanalyzed = "No" AND cohortYear >= 1997
    AND validity = "crude" AND vaccine IN vaxs
}
