% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{surviving}
\alias{surviving}
\title{Obtain statistics on child population}
\usage{
surviving(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Newborn children by year
}
\description{
Obtain statistics on child population
}
\details{
SELECT annum, valor
  FROM DEMOGRAPHIC
  WHERE country = CCODE AND annum >= 1997 AND indicator = 'SI' AND valor > 0
}
