% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{svy.scope}
\alias{svy.scope}
\title{svy.scope: When checking for svy.thrs, use surveys across +/- 2 years}
\usage{
svy.scope()
}
\value{
2
}
\description{
svy.scope: When checking for svy.thrs, use surveys across +/- 2 years
}
