% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{svy.thrs}
\alias{svy.thrs}
\title{svy.thrs: differences between reported coverage and survey coverage of
more than 10 percentage points raise flags in GoC and anchor points}
\usage{
svy.thrs()
}
\value{
10
}
\description{
svy.thrs: differences between reported coverage and survey coverage of
more than 10 percentage points raise flags in GoC and anchor points
}
