% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{target}
\alias{target}
\title{Obtain statistics on vaccinated children}
\usage{
target(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with target numbers
}
\description{
Obtain statistics on vaccinated children
}
\details{
SELECT vaccine, annum, reportedDenom, reportedNum, coverage
  FROM REPORTED_NUMERATOR_DENOMINATOR
  WHERE country = CCODE AND reportedNum > 0
}
