% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tround.R
\name{tround}
\alias{tround}
\title{Simulate Prolog's "outward" rounding}
\usage{
tround(number)
}
\arguments{
\item{number}{a float}
}
\value{
rounded toward next even number
}
\description{
Simulate Prolog's "outward" rounding
}
