% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{unpd.thrs}
\alias{unpd.thrs}
\title{unpd.thrs: differences between reported coverage and UNPD data of
more than 10 percentage points raise flags in GoC}
\usage{
unpd.thrs()
}
\value{
10
}
\description{
unpd.thrs: differences between reported coverage and UNPD data of
more than 10 percentage points raise flags in GoC
}
