% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{vaccinated}
\alias{vaccinated}
\title{Obtain statistics on vaccinated children}
\usage{
vaccinated(mdb, ccode)
}
\arguments{
\item{mdb}{path to database}

\item{ccode}{Country code, e.g. afg}
}
\value{
Year by Vaccine table with numbers
}
\description{
Obtain statistics on vaccinated children
}
\details{
SELECT vaccine, annum, reportedDenom, reportedNum, coverage
 FROM REPORTED_NUMERATOR_DENOMINATOR
 WHERE country = CCODE AND reportedNum > 0
}
