% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{Vn}
\alias{Vn}
\title{List of all vaccine names}
\usage{
Vn()
}
\value{
returns of a list with the abbreviated names of the vaccines used in the
Wuenic program
}
\description{
List of all vaccine names
}
\details{
This function will be removed at some 
point (replaced by a data-driven global variable)
}
