% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_anchor.R
\name{wuenic.anchor}
\alias{wuenic.anchor}
\title{Integrate estimates from reported data and surveys}
\usage{
wuenic.anchor(TS.Cov, TS.Src, Svy.Cov, Svy.Expl, Decisions)
}
\arguments{
\item{TS.Cov}{Coverage from time series}

\item{TS.Src}{Source of the coverage from time series}

\item{Svy.Cov}{Coverage from surveys}

\item{Svy.Expl}{Explanation derived from surveys}

\item{Decisions}{List of work group decisions}
}
\value{
List with Coverage, Rule/Source, and some explanation (info)
}
\description{
Define the according anchor points (reported | survey | wgd)
}
