% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_calibrate.R
\name{wuenic.calibrate}
\alias{wuenic.calibrate}
\title{Estimate coverage if information is incomplete}
\usage{
wuenic.calibrate(
  TS.Cov,
  TS.Src,
  Anchor.Cov,
  Anchor.Rule,
  Anchor.Info,
  Decisions
)
}
\arguments{
\item{TS.Cov}{coverage from time series}

\item{TS.Src}{source of coverage from time series}

\item{Anchor.Cov}{Coverage at anchor points (with converging coverage estimates)}

\item{Anchor.Rule}{Rules for anchor points}

\item{Anchor.Info}{Additional information at anchor points}

\item{Decisions}{List of work group decisions}
}
\value{
List with coverage and source, and some explanation
}
\description{
Distinguish different cases
}
\details{
* at anchor point
* between anchor points
* before anchor point
* after anchor point
* no anchor points
}
