% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_check.R
\name{wuenic.check}
\alias{wuenic.check}
\title{Check reported data}
\usage{
wuenic.check(Cov, Decisions)
}
\arguments{
\item{Cov}{coverage from admin and gov data}

\item{Decisions}{List of work group decisions}
}
\value{
Matrix with Flags and Explanation
}
\description{
Reasons to exclude reported data are working group decisions,
coverage > 100% or temporal inconsistency (sudden jumps, sudden declines, 
etc.)
}
