% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_explanations.R
\name{wuenic.explanations}
\alias{wuenic.explanations}
\title{Collect explanations in natural language terms}
\usage{
wuenic.explanations(
  Rep.Expl,
  Svy.Excl,
  Svy.Expl.Acc,
  Rej.Info,
  Decisions,
  Info,
  Change,
  GoC.Expl
)
}
\arguments{
\item{Rep.Expl}{Explanations for reported data}

\item{Svy.Excl}{Excluded surveys}

\item{Svy.Expl.Acc}{Accepted surveys}

\item{Rej.Info}{Explanations if government data has been rejected}

\item{Decisions}{List of work group decisions}

\item{Info}{General information}

\item{Change}{Changes from legacy estimates}

\item{GoC.Expl}{Explanations of confidence grading}
}
\value{
Matrix with explanations
}
\description{
# Prolog
collect_explanations(C, V, Y, Explanations) :-
    findall(Expl, explanation(C, V, Y, Expl), Explanations).
}
