% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_goc.R
\name{wuenic.goc}
\alias{wuenic.goc}
\title{Grade of confidence}
\usage{
wuenic.goc(
  Ereq,
  Rep.Cov,
  Svy.Cov,
  Cov,
  Rule,
  Rub,
  Decisions,
  Vaccinated,
  Births,
  Surviving
)
}
\arguments{
\item{Ereq}{Is an estimate required?}

\item{Rep.Cov}{Reported coverage}

\item{Svy.Cov}{Coverage in surveys}

\item{Cov}{Coverage synthesis}

\item{Rule}{Decision rule for coverage estimate}

\item{Rub}{Rubella vaccine given in combination with other vaccine}

\item{Decisions}{List of work group decisions}

\item{Vaccinated}{General information}

\item{Surviving}{Birth statistics}

\item{Birth}{Birth statistics}
}
\value{
Matrix with grade of confidence and explanation
}
\description{
Do reported data and surveys provide similar estimates, consistent with
child statistics?
}
