% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_legacy.R
\name{wuenic.legacy}
\alias{wuenic.legacy}
\title{Flag modifications in the program code that change the coverage estimates}
\usage{
wuenic.legacy(Bounded, Legacy)
}
\arguments{
\item{Bounded}{Current estimate}

\item{Legacy}{Previous estimate}
}
\value{
Matrix with comments
}
\description{
Flag modifications in the program code that change the coverage estimates
}
