% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load.R
\name{wuenic.load}
\alias{wuenic.load}
\title{Read date from pl-file}
\usage{
wuenic.load(fname)
}
\arguments{
\item{fname}{Prolog file with country-specific information}
}
\description{
Read date from pl-file
}
\details{
Legacy code, for compatibility with V4R. Will be removed soon.
}
