% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_reported.R
\name{wuenic.reported}
\alias{wuenic.reported}
\title{Add information from government and administration}
\usage{
wuenic.reported(Admin, Gov, Decisions)
}
\arguments{
\item{Admin}{admin data}

\item{Gov}{governemt data}

\item{Decisions}{List of work group decisions}
}
\value{
List with coverage and source, and some explanation
}
\description{
Reported to WHO and UNICEF is government estimate. If government estimate
missing, then reported is administrative data. Working group decisions may
override this rule.
}
\details{
In the following implementation, we start at the lowest level, which is the
administrative data. If the government also reports the coverage, the former
numbers are overwritten.
}
