% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_survey.R
\name{wuenic.survey}
\alias{wuenic.survey}
\title{Analyze survey data.}
\usage{
wuenic.survey(Survey, Decisions)
}
\arguments{
\item{Survey}{Long table with survey information}

\item{Decisions}{List of work group decisions}
}
\value{
List with coverage, Explanations, and information on accepted and excluded
surveys
}
\description{
This is a bit more complicated, since we have multiple surveys for the
same year. Therefore, the survey results are stored in a three-dimensional
array, Vaccine x Year x Survey Id.
}
\details{
Analysis is done in several steps
- select subset of surveys that apply for child diseases
- remove surveys with low information (e.g. sample size < 300)
- correct "recall bias" in surveys, e.g. when 3rd shot rate is implausibly
  high
- remove surveys when working group decisions say so
- average the available surveys for a given vaccine
}
