% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_toplevel.R
\name{wuenic.top}
\alias{wuenic.top}
\title{Estimate coverage based on all information}
\usage{
wuenic.top(Ereq, Rule, Info, Cov, Rub, firstRubellaAtSecondMCV, Decisions)
}
\arguments{
\item{Ereq}{Matrix with booleans, is an estimate required?}

\item{Rule}{Decision rule (accept reported etc.)}

\item{Info}{Additional information}

\item{Cov}{Coverage estimate}

\item{Rub}{Where to obtain the data for the rubella vaccination}

\item{firstRubellaAtSecondMCV}{Where to obtain the data for the rubella vaccination (default)}

\item{Decisions}{List of work group decisions}
}
\value{
List with coverage (incl. bounded 0...99) and source, and some explanation
}
\description{
Distinguish different cases
}
\details{
* at anchor point
* between anchor points
* before anchor point
* after anchor point
* no anchor points
}
