% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_ts.R
\name{wuenic.ts}
\alias{wuenic.ts}
\title{Generate time series based on reported coverage}
\usage{
wuenic.ts(Ereq, Rep.Cov, Rep.Src, Reject)
}
\arguments{
\item{Ereq}{if an estimate is required}

\item{Rep.Cov}{reported coverage}

\item{Rep.Src}{Source of reported coverage}

\item{Reject}{Information on decisions (reject reported)}
}
\value{
list with coverage and source
}
\description{
Use interpolation and extrapolation for years without reported coverage.
}
