% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{Yn}
\alias{Yn}
\title{Year range of Wuenic program}
\usage{
Yn()
}
\value{
1985:2023
}
\description{
Year range of Wuenic program
}
\details{
This function will be removed at some 
point (replaced by a data-driven global variable)
}
