% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{YV.char}
\alias{YV.char}
\alias{YV.int}
\alias{YV.bool}
\alias{YV.real}
\title{YV.char, YV.int, YV.bool: Year x Vaccine matrices}
\usage{
YV.char(fill = NA_character_)

YV.int(fill = NA_integer_)

YV.bool(fill = FALSE)

YV.real(fill = NA_real_)
}
\arguments{
\item{fill}{(default: NA) value(s) to fill the matrix}
}
\value{
Year x Vaccine matrices
}
\description{
YV.char, YV.int, YV.bool: Year x Vaccine matrices
}
